misc: Move NmsUtils#getSyncableAttributeListFor to AttributeUtils#syncableAttributesFor
